spaceport/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.15
fastapi==0.115.6
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
opentelemetry-api==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-openai==0.34.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-semantic-conventions-ai==0.4.2
packaging==24.2
pydantic==2.10.3
pydantic_core==2.27.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
tiktoken==0.8.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
wrapt==1.17.0
zipp==3.21.0
